update store table

+++ src/main/java/api/lemonico/entity/AuthorityEntity.java
@@ -47,5 +47,5 @@ public class AuthorityEntity extends LcEntity
     LocalDateTime modifiedAt;
     /** 削除フラグ（0: 未削除 1: 削除済） */
     @Column(name = "is_deleted")
-    Byte isDeleted;
+    Integer isDeleted;
 }
