+++ src/main/java/api/lemonico/entity/WarehouseStoreEntity.java
@@ -50,5 +50,5 @@ public class WarehouseStoreEntity extends LcEntity
     LocalDateTime modifiedAt;
     /** 削除フラグ（0: 未削除 1: 削除済） */
     @Column(name = "is_deleted")
-    Byte isDeleted;
+    Integer isDeleted;
 }
